backend/classification_model/requirements.txt
tensorflow
numpy
opencv-python
scikit-learn
matplotlib
Pillow
h5py

